blog/requirement.txt
django
markdown
django-markdownx
ckeditor

